requirements.txt
agno==2.5.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bitarray==3.8.0
certifi==2026.2.25
click==8.3.1
distro==1.9.0
docstring-parser==0.17.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.24.0
packaging==26.0
pdfreader==0.1.15
pillow==12.1.1
pip==26.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.2
pypdf==6.7.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pyyaml==6.0.3
rich==14.3.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
